cleanup lets_split
