+++ keyboards/lets_split/keymaps/OLED_sample/config.h
@@ -1,23 +1,18 @@
-/*
-Copyright 2012 Jun Wako <wakojun@gmail.com>
+#ifndef CONFIG_USER_H
+#define CONFIG_USER_H
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
+#include "../../config.h"
 
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
+/* Use I2C or Serial, not both */
 
 #define USE_I2C
 //#define USE_SERIAL
-//#define MASTER_RIGHT
+
+/* Select hand configuration */
+
+#define MASTER_LEFT
+// #define _MASTER_RIGHT
+// #define EE_HANDS
 #define FLIP_HALF
 
 #define SSD1306OLED
@@ -40,3 +35,5 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define RGBLIGHT_HUE_STEP 10
 #define RGBLIGHT_SAT_STEP 17
 #define RGBLIGHT_VAL_STEP 17
+
+#endif
