+++ keyboards/lets_split/keymaps/adam/config.h
@@ -1,39 +1,24 @@
-/*
-Copyright 2012 Jun Wako <wakojun@gmail.com>
+#ifndef CONFIG_USER_H
+#define CONFIG_USER_H
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
+#include "../../config.h"
 
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
+/* Use I2C or Serial, not both */
 
 #define USE_SERIAL
+// #define USE_I2C
 
-//#define MASTER_LEFT
-// #define MASTER_RIGHT
+/* Select hand configuration */
+
+// #define MASTER_LEFT
+// #define _MASTER_RIGHT
 #define EE_HANDS
 
-#ifdef SUBPROJECT_rev1
-    #include "../../rev1/config.h"
-#endif
 #ifdef SUBPROJECT_rev2
-    #include "../../rev2/config.h"
     /* RGB Underglow */
     #undef RGBLED_NUM
     #define RGBLIGHT_ANIMATIONS
     #define RGBLED_NUM 8
-#endif
-#ifdef SUBPROJECT_rev2fliphalf
-	#include "../../rev2fliphalf/config.h"
 #endif
 
 #undef TAPPING_TERM
@@ -41,3 +26,5 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define PREVENT_STUCK_MODIFIERS
 #define IGNORE_MOD_TAP_INTERRUPT
 #define PERMISSIVE_HOLD
+
+#endif
