+++ keyboards/lets_split/keymaps/default/Makefile
@@ -1,0 +1,3 @@
+ifndef QUANTUM_DIR
+	include ../../../../Makefile
+endif
