+++ keyboards/lets_split/keymaps/khord/config.h
@@ -1,33 +1,21 @@
-/*
-Copyright 2012 Jun Wako <wakojun@gmail.com>
+#ifndef CONFIG_USER_H
+#define CONFIG_USER_H
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
+#include "../../config.h"
 
 #define TAPPING_TERM 150
 
+/* Use I2C or Serial, not both */
+
 #define USE_SERIAL
+// #define USE_I2C
+
+/* Select hand configuration */
+
+//#define MASTER_LEFT
+// #define _MASTER_RIGHT
 #define EE_HANDS
 
 #define RGBLIGHT_ANIMATIONS
 
-#ifdef SUBPROJECT_rev1
-    #include "../../rev1/config.h"
-#endif
-#ifdef SUBPROJECT_rev2
-    #include "../../rev2/config.h"
-#endif
-#ifdef SUBPROJECT_rev2fliphalf
-	#include "../../rev2fliphalf/config.h"
 #endif
