+++ keyboards/orthodox/keymaps/default/Makefile
@@ -1,0 +1,3 @@
+ifndef QUANTUM_DIR
+	include ../../../../Makefile
+endif
