+++ keyboards/orthodox/keymaps/default/config.h
@@ -10,16 +10,8 @@
 
 /* Select hand configuration */
 
-#define MASTER_LEFT
+// #define MASTER_LEFT
 // #define _MASTER_RIGHT
-// #define EE_HANDS
-
-
-#ifdef SUBPROJECT_rev2
-    /* RGB Underglow */
-    #undef RGBLED_NUM
-    #define RGBLIGHT_ANIMATIONS
-    #define RGBLED_NUM 8
-#endif
+#define EE_HANDS
 
 #endif
